fixed registration flow
